Add activity idea to 4.3 Censor transactions
The idea behind this activity is to show that the transaction pool operates off chain and there is no consensus over it. That means that an individual node can run alternate code in the transaction pool while still producing valid blocks and verifying others' blocks correctly.

As a bonus, the code that needs to be changed is in the runtime (block builder api) which foreshadows the runtime apis discussed in section 4.4.