Don't recommend Jupyter, Clean up templates (#361)
This PR removes some of the templates (the ones that were basically
empty) as well as all the jupyter examples.

It also entirely removes the recommendation to use jupyter and EvCxR.
This does not forbid using it. Instructors who like it (if there are
any) are welcome to use it just as they are welcome to use any other
tool. In fact, there are still two jupyter notebooks in the repo, which
is fine.

@NukeManDan can you please take a look at the remaining templates from
the previous wave? I have a sense that they can be combined or distilled
down so that new contributors have less to read before starting to
contributw.