Module 3: Blockchain - Restructure for Wave 2 (#360)
This PR restructures the Blockchain module for the upcoming Buenos Aires
cohort. Not all of the content is created yet, and this PR does not need
to be delayed waiting for it all to be created. This PR can be merged
when we've reached consensus about the lesson plan that is in
`syllabus/3-Blockchain/README.md`.

The major content changes are:
* Add lecture about P2P networks
* Add lecture about light clients
* Split intro lecture into two smaller lectures and break it up with BFS
coding activity
* Add details about Grandpa finality gadget (previously in module 4) and
accompanying board game
* Remove the first, too-elementary activity about drawing a blockchain.

There is also some restructuring of the directory. Basically I've just
made it flatter and thus easier to navigate.

Co-authored-by: Dan Shields <35669742+NukeManDan@users.noreply.github.com>