Refactor the simulator in XCM to be common for multiple simulator exercises (#372)
Changed up directory structure and format of initialization of the
simulator for the various simulator exercises.
The goal when creating new simulator exercises is to just create a few
tests importing this new `xcm-simulator-for-exercises` crate. This way
we can all customize the xcm-simulator as we see fit for students to use
during various constructed exercises.