Review encryption slides
The main changes here are being more explicit about APIs, and
emphasizing AE(AD) and hybrid encryption, because those are the
primitives that are actually used, and it is very easy to make a mistake
when using standard encryption. Also, most encryption libraries that are
sufficiently high level only offer aead, as it is safest.